backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ImageIO==2.37.2
lazy_loader==0.4
networkx==3.6.1
numpy==2.4.1
opencv-python==4.13.0.90
packaging==26.0
pillow==12.1.0
PyJWT==2.10.1
scikit-image==0.26.0
scipy==1.17.0
sqlparse==0.5.5
tifffile==2026.1.14
